haineng 提交修改代码

+++ customer-convert-server/src/main/java/com/smart/sso/server/model/dto/CustomerFeatureResponse.java
@@ -78,7 +78,7 @@ public class CustomerFeatureResponse {
         // 交易风格
         private Feature.CustomerConclusion tradingStyle = new Feature.CustomerConclusion();
         // 客户的回复频次
-        private Integer customerResponse;
+        private Float customerResponse;
         // 是否购买类似的产品
         private Feature.CustomerConclusion purchaseSimilarProduct = new Feature.CustomerConclusion();
     }
